backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI and ML
openai==1.3.7
langchain==0.0.340
langchain-openai==0.0.2
pinecone-client==2.2.4
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
clip-by-openai==1.0
sentence-transformers==2.2.2

# Computer Vision
opencv-python==4.8.1.78
Pillow==10.1.0
mediapipe==0.10.8

# Data processing
numpy==1.24.4
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP and API
httpx==0.25.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring and logging
structlog==23.2.0

# Social media and web scraping
beautifulsoup4==4.12.2
requests==2.31.0

# Image processing and storage
boto3==1.34.0
google-cloud-vision==3.4.5

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0